requirements.txt
opencv-python
matplotlib
deepface
streamlit
numpy
requests
Pillow
seaborn
mtcnn